Move stub class into a proper class.
