+++ src/main/java/seedu/todo/ui/views/HelpView.java
@@ -7,6 +7,7 @@ import javafx.fxml.FXML;
 import javafx.scene.layout.Pane;
 import javafx.stage.Stage;
 import seedu.todo.commons.util.FxViewUtil;
+import seedu.todo.controllers.CommandDefinition;
 import seedu.todo.ui.UiPartLoader;
 import seedu.todo.ui.components.HelpCommandItem;
 
@@ -15,7 +16,7 @@ public class HelpView extends View {
     private static final String FXML_PATH = "views/HelpView.fxml";
     
     // Props
-    public List<CommandDefinitionStub> commandDefinitions = new ArrayList<CommandDefinitionStub>();
+    public List<CommandDefinition> commandDefinitions = new ArrayList<CommandDefinition>();
     
     // FXML
     @FXML
@@ -39,7 +40,7 @@ public class HelpView extends View {
         HelpCommandItem.reset(helpCommandsPlaceholder);
         
         // Load help commands
-        for (CommandDefinitionStub command : commandDefinitions) {
+        for (CommandDefinition command : commandDefinitions) {
             HelpCommandItem item = HelpCommandItem.load(primaryStage, helpCommandsPlaceholder);
             item.commandName = command.getCommandName();
             item.commandDescription = command.getCommandDescription();
@@ -48,43 +49,5 @@ public class HelpView extends View {
         }
     }
     
-    // TODO: Replace stub
-    public static class CommandDefinitionStub {
-        
-        private String commandName;
-        private String commandDescription;
-        private String commandSyntax;
-        
-        // TODO: Remove after testing
-        public CommandDefinitionStub(String name, String desc, String syntax) {
-            commandName = name;
-            commandDescription = desc;
-            commandSyntax = syntax;
-        }
-        
-        public String getCommandName() {
-            return commandName;
-        }
-        
-        public void setCommandName(String commandName) {
-            this.commandName = commandName;
-        }
-        
-        public String getCommandDescription() {
-            return commandDescription;
-        }
-        
-        public void setCommandDescription(String commandDescription) {
-            this.commandDescription = commandDescription;
-        }
-        
-        public String getCommandSyntax() {
-            return commandSyntax;
-        }
-        
-        public void setCommandSyntax(String commandSyntax) {
-            this.commandSyntax = commandSyntax;
-        }
-    }
 
 }
